added unit tests arounf ffoi station and station forcast model
